fix IllegalArgumentException from ByteBuffer.limit() when using the ChunkDecoder
git-svn-id: https://svn.apache.org/repos/asf/httpcomponents/httpcore/trunk@781964 13f79535-47bb-0310-9956-ffa450edef68

+++ httpcore-nio/src/examples/org/apache/http/examples/nio/NHttpFileServer.java
@@ -217,7 +217,7 @@ public class NHttpFileServer {
                         fileChannel, idx, Long.MAX_VALUE);
             } else {
                 transferred = fileChannel.transferFrom(
-                        new ContentDecoderChannel(decoder), idx, Long.MAX_VALUE);
+                        new ContentDecoderChannel(decoder), idx, Integer.MAX_VALUE);
             }
 
             if(transferred > 0)
